// api/test/src/product/dto/update-product.dto.ts
import { PartialType } from '@nestjs/swagger';
import { CreateProductDto } from './create-product.dto';

export class UpdateProductDto extends PartialType(CreateProductDto) {
    name?: string;
    description?: string;
    price?: number;
    quantity?: number;
    variationOptions?: {
      quantity: number; 
    }[]; 
}